specific item discounts implemented

+++ discount/models/structDiscount.go
@@ -3,6 +3,6 @@ package models
 type Discount struct {
 	Id     int    `json:"id,omitempty"`
 	Sku     string    `json:"sku,omitempty"`
-	DiscountPercent     int    `json:"discountPercent,omitempty"`
-	DiscountAmount     int    `json:"discountAmount,omitempty"`
+	DiscountPercent     float64    `json:"discountPercent,omitempty"`
+	DiscountAmount     float64    `json:"discountAmount,omitempty"`
 }
